requirements.txt
fastapi
uvicorn
python-binance
requests
transformers
torch
apscheduler
jinja2
